Fix test failures in Windows and add negative tests
