+++ compiler-tools/benchmark/runner.go
@@ -70,11 +70,11 @@ func (b *benchmark) run() error {
 	defer b.removeWorktree(headWorktree)
 
 	fmt.Printf("Building interpreter for %s...\n", b.baseRef)
-	if err := b.buildInterpreter(baseWorktree, b.baseRef, builtInterpreter); err != nil {
+	if err := b.buildInterpreter(baseWorktree, b.baseRef, builtInterpreterBinaryName()); err != nil {
 		return err
 	}
 	fmt.Printf("Building interpreter for %s...\n", b.headRef)
-	if err := b.buildInterpreter(headWorktree, b.headRef, builtInterpreter); err != nil {
+	if err := b.buildInterpreter(headWorktree, b.headRef, builtInterpreterBinaryName()); err != nil {
 		return err
 	}
 
@@ -84,7 +84,7 @@ func (b *benchmark) run() error {
 	}
 	defer func() { _ = os.RemoveAll(exportDir) }()
 
-	var results []runResult
+	results := make([]runResult, 0, len(target.paths))
 	for _, path := range target.paths {
 		cmds := b.benchmarkCmdPair(baseWorktree, headWorktree, target.root, path, target.mode)
 		exportPath := filepath.Join(exportDir, fmt.Sprintf("%s.json", sanitize(path)))
@@ -92,15 +92,8 @@ func (b *benchmark) run() error {
 		if err != nil {
 			return err
 		}
-		label := target.label
-		if target.mode == multipleFilesMode {
-			label = path
-			for strings.HasPrefix(label, "../") {
-				label = strings.TrimPrefix(label, "../")
-			}
-		}
 		results = append(results, runResult{
-			label:  label,
+			label:  benchmarkResultLabel(target, path),
 			export: *export,
 		})
 	}
@@ -111,12 +104,10 @@ func (b *benchmark) run() error {
 		Generated: time.Now(),
 		results:   results,
 	}
-	if b.config.exportPath != "" {
-		if err := rep.export(b.config.exportPath); err != nil {
-			return err
-		}
-		fmt.Printf("Benchmark report exported to %s\n", b.config.exportPath)
+	if err := rep.export(b.config.exportPath); err != nil {
+		return err
 	}
+	fmt.Printf("Benchmark report exported to %s\n", b.config.exportPath)
 	return nil
 }
 
@@ -161,7 +152,7 @@ func (b *benchmark) runHyperfine(cmds []string, jsonExportPath string) (*benchEx
 }
 
 func (b *benchmark) benchmarkCmdArgs(ref, interpreter, root, target string, mode targetMode) []string {
-	command := fmt.Sprintf("%s run %s", shellQuote(interpreter), shellQuote(target))
+	command := formatBenchmarkCommand(interpreter, target)
 	if mode == multipleFilesMode {
 		ref = fmt.Sprintf("%s (%s)", ref, getRelativeLabel(root, target))
 	}
@@ -169,8 +160,8 @@ func (b *benchmark) benchmarkCmdArgs(ref, interpreter, root, target string, mode
 }
 
 func (b *benchmark) benchmarkCmdPair(baseWorktree, headWorktree, root, target string, mode targetMode) []string {
-	baseCmd := b.benchmarkCmdArgs(b.baseRef, filepath.Join(baseWorktree, builtInterpreter), root, target, mode)
-	headCmd := b.benchmarkCmdArgs(b.headRef, filepath.Join(headWorktree, builtInterpreter), root, target, mode)
+	baseCmd := b.benchmarkCmdArgs(b.baseRef, filepath.Join(baseWorktree, builtInterpreterBinaryName()), root, target, mode)
+	headCmd := b.benchmarkCmdArgs(b.headRef, filepath.Join(headWorktree, builtInterpreterBinaryName()), root, target, mode)
 	return append(baseCmd, headCmd...)
 }
 
@@ -195,10 +186,44 @@ func sanitize(ref string) string {
 }
 
 func shellQuote(s string) string {
-	if runtime.GOOS == "windows" {
+	return shellQuoteForOS(runtime.GOOS, s)
+}
+
+func shellQuoteForOS(goos, s string) string {
+	if goos == "windows" {
 		return `"` + strings.ReplaceAll(s, `"`, `\"`) + `"`
 	}
 	return "'" + strings.ReplaceAll(s, "'", `'"'"'`) + "'"
+}
+
+func formatBenchmarkCommand(interpreter, target string) string {
+	return formatBenchmarkCommandForOS(runtime.GOOS, interpreter, target)
+}
+
+func formatBenchmarkCommandForOS(goos, interpreter, target string) string {
+	cmd := fmt.Sprintf("%s run %s", shellQuoteForOS(goos, interpreter), shellQuoteForOS(goos, target))
+	if goos == "windows" {
+		return `"` + cmd + `"`
+	}
+	return cmd
+}
+
+func builtInterpreterBinaryName() string {
+	if runtime.GOOS == "windows" {
+		return builtInterpreter + ".exe"
+	}
+	return builtInterpreter
+}
+
+func benchmarkResultLabel(target *benchmarkTarget, path string) string {
+	if target.mode != multipleFilesMode {
+		return target.label
+	}
+	label := path
+	for strings.HasPrefix(label, "../") {
+		label = strings.TrimPrefix(label, "../")
+	}
+	return label
 }
 
 func runCmd(dir, name string, args ...string) error {
