+++ compiler-tools/benchmark/testdata/project/Ballerina.toml
@@ -1,0 +1,4 @@
+[package]
+org = "testorg"
+name = "bench_pkg"
+version = "0.1.0"
